Ergasia6
This  code read from the user the number of the month, the year and the day of the week exp. for Sunday tap 1,for Tuesday tap 3. After that the code prints on the screen the name of the month, the year in integer and the calendar of this month only.